refactor(audit): move _resolve_client_ip to core/audit.py
Pure code move in preparation for IP anonymization (6C) and
audit-trail extension to chat/admin_replay paths (6B).

No behavior change.

Refs: KIS-AUDIT-6C-IP-PRIVACY
